backend/requirements.txt
flask==2.2.2
flask-cors==4.0.0
requests==2.31.0
werkzeug==2.2.2
python-dotenv==1.0.0
azure-identity==1.12.0
azure-keyvault-secrets==4.8.0


